@@ -39,72 +33,88 @@
       "json",
       "html"
     ],
-    "all": true
+    "sourceMap": true,
+    "instrument": true
   },
   "dependencies": {
-    "@types/chai-subset": "^1.3.0",
+    "@json-api/query-parser": "^1.2.0",
     "@types/debug": "0.0.29",
-    "@types/express": "^4.0.36",
-    "@types/immutable": "^3.8.7",
-    "@types/mongodb": "^2.2.7",
-    "@types/mongoose": "^4.7.18",
+    "@types/depd": "^1.1.32",
+    "@types/express": "^4.11.1",
+    "@types/lodash": "^4.14.107",
+    "@types/node": "^8.10.9",
     "@types/pluralize": "0.0.27",
-    "@types/q": "^1.0.2",
-    "@types/qs": "^6.5.0",
-    "@types/ramda": "^0.24.4",
-    "@types/raw-body": "^2.1.4",
-    "@types/url-template": "^2.0.27",
-    "content-type": "1.x.x",
+    "@types/pug": "^2.0.4",
+    "@types/qs": "^6.5.1",
+    "@types/ramda": "^0.25.24",
+    "@types/url-template": "^2.0.28",
+    "content-type": "^1.0.4",
     "dasherize": "2.0.x",
-    "debug": "^2.6.8",
+    "debug": "^3.1.0",
+    "depd": "^1.1.2",
     "flat": "^1.2.1",
-    "immutable": "^3.7.5",
-    "jade": "1.11.x",
-    "lodash": "^3.10.0",
+    "immutable": "^3.8.2",
+    "lodash": "^4.17.5",
     "negotiator": "github:ethanresnick/negotiator#full-parse-access",
     "pluralize": "0.0.11",
-    "q": "1.4.1",
-    "qs": "^6.5.0",
-    "ramda": "^0.24.1",
-    "raw-body": "1.3.x",
+    "pug": "^2.0.3",
+    "qs": "^6.5.1",
+    "ramda": "^0.25.0",
+    "raw-body": "2.x.x",
+    "supports-color": "^4.5.0",
     "url-template": "^2.0.4",
-    "vary": "^1.1.1"
+    "vary": "^1.1.2"
   },
   "devDependencies": {
-    "@ethanresnick/tslint-config": "^5.5.1",
-    "@types/mocha": "^2.2.41",
-    "chai": "^1.9.2",
-    "chai-subset": "^1.5.0",
-    "coveralls": "^2.13.1",
+    "@ethanresnick/tslint-config": "^5.9.1",
+    "@types/chai-subset": "^1.3.1",
+    "@types/mocha": "^2.2.48",
+    "@types/mongodb": "^2.2.20",
+    "@types/mongoose": "^4.7.36",
+    "@types/mongoose-geojson-schema": "^2.1.1",
+    "@types/sinon": "^4.3.1",
+    "@types/superagent": "^3.5.8",
+    "body-parser": "^1.18.2",
+    "chai": "^4.1.2",
+    "chai-subset": "^1.6.0",
+    "coveralls": "^2.13.3",
     "cpr": "^2.2.0",
     "eslint": "0.x.x",
-    "express": "^4.15.3",
+    "express": "^4.16.3",
     "gulp": "^3.8.6",
     "istanbul": "0.3.x",
     "mocha": "^2.5.3",
-    "mongoose": "4.0.x",
+    "mongoose": "^4.13.12",
+    "mongoose-geojson-schema": "^2.1.2",
     "node-mongoose-fixtures": "^0.2.4",
-    "nyc": "^11.0.3",
-    "rimraf": "^2.6.1",
-    "sinon": "1.10.2",
-    "superagent": "1.2.x",
-    "ts-node": "^3.2.0",
-    "tslint": "^5.5.0",
-    "typescript": "^2.4.1"
+    "nyc": "^11.7.1",
+    "rimraf": "^2.6.2",
+    "sinon": "^4.5.0",
+    "source-map-support": "^0.5.4",
+    "superagent": "github:visionmedia/superagent",
+    "ts-node": "^3.3.0",
+    "tslint": "^5.9.1",
+    "typescript": "^2.8.3"
   },
   "peerDependencies": {
-    "mongoose": "^4.0.0",
+    "mongoose": "^4.7.0",
     "express": "^4.0.0"
   },
   "scripts": {
     "clean": "rimraf build",
-    "lint": "tslint --project tsconfig.json --format verbose --config node_modules/@ethanresnick/tslint-config/tslint.json",
-    "test": "export NODE_ENV=testing; mocha --compilers ts:ts-node/register --recursive test/unit/ test/integration/",
+    "lint": "npm run lint:src & npm run lint:test",
+    "lint:src": "tslint --project tsconfig.json --format verbose --exclude 'test/**/*' --config ./src/tslint.json",
+    "lint:test": "tslint --project tsconfig.json --format verbose --exclude 'src/**/*' --config ./test/tslint.json",
+    "test": "npm run test:install-fixtures && npm run test:fixtures-installed",
+    "test:install-fixtures": "NODE_ENV=testing ts-node test/app/database/install-fixtures.ts",
+    "test:fixtures-installed": "NODE_ENV=testing mocha --compilers ts:ts-node/register --recursive test/unit/ test/integration/ --full-trace --check-leaks",
+    "test:debug": "npm run test -- --inspect-brk",
     "build:lintfree": "npm run clean && cpr templates build/templates && echo Using TypeScript && tsc --version && tsc --pretty",
-    "build": "npm run build:lintfree && npm run lint",
-    "prepublish": "npm run test && npm run build:lintfree",
-    "cover": "NODE_ENV=test nyc npm t",
-    "cover:view": "npm run cover && open coverage/lcov-report/index.html",
+    "build": "npm run lint && npm run build:lintfree",
+    "prepare": "npm run build",
+    "prepublishOnly": "npm run test",
+    "cover": "NODE_ENV=test npm run test:install-fixtures && nyc npm run test:fixtures-installed",
+    "cover:view": "npm run cover && open coverage/index.html",
     "watch": "npm run build -- --watch",
     "watch:lintfree": "npm run build:lintfree -- --watch",
     "watch:test": "npm run test -- --watch"
