@@ -1,76 +1,168 @@
-import Q = require("q");
+import R = require("ramda");
+// Import mongo just for one of its type declarations (not imported at runtime).
+// tslint:disable-next-line no-implicit-dependencies
 import mongodb = require("mongodb");
 import mongoose = require("mongoose");
 import pluralize = require("pluralize");
 
-import {arrayContains, arrayValuesMatch} from "../../util/arrays";
-import {deleteNested} from "../../util/misc";
-import {forEachArrayOrVal, mapResources, groupResourcesByType} from "../../util/type-handling";
+import { Model, Document } from "mongoose";
+
+import { AndExpression, SupportedOperators, ExpressionSort } from "../../types/";
+import { isId as isIdentifier } from '../../steps/pre-query/parse-query-params';
+import { partition, setDifference, reduceToObject } from "../../util/misc";
+import { values as objectValues } from "../../util/objectValueEntries";
+import {
+  getReferencePaths, getReferencedModelName,
+  getDiscriminatorKey, getVersionKey, getMetaKeys
+} from './utils/schema';
+import { getTypePath } from "./utils/subtyping";
+import docToResource from "./utils/doc-to-resource";
+import { getTypeName } from "../../util/naming-conventions";
 import * as util from "./lib";
-import Resource, { ResourceWithId } from "../../types/Resource";
-import Collection from "../../types/Collection";
-import Linkage from "../../types/Linkage";
-import Relationship from "../../types/Relationship";
-import APIError from "../../types/APIError";
+import * as Errors from '../../util/errors';
+import Data from "../../types/Generic/Data";
+import Resource, { ResourceWithTypePath } from "../../types/Resource";
+import ResourceIdentifier from "../../types/ResourceIdentifier";
 import FieldDocumentation from "../../types/Documentation/Field";
 import FieldTypeDocumentation from "../../types/Documentation/FieldType";
 import RelationshipTypeDocumentation from "../../types/Documentation/RelationshipType";
-import { Adapter } from '../AdapterInterface';
+import { Adapter, TypeInfo, TypeIdMapOf } from "../AdapterInterface";
+import CreateQuery from "../../types/Query/CreateQuery";
+import FindQuery from "../../types/Query/FindQuery";
+import DeleteQuery from "../../types/Query/DeleteQuery";
+import UpdateQuery from "../../types/Query/UpdateQuery";
+import AddToRelationshipQuery from "../../types/Query/AddToRelationshipQuery";
+import RemoveFromRelationshipQuery from "../../types/Query/RemoveFromRelationshipQuery";
 
 export default class MongooseAdapter implements Adapter<typeof MongooseAdapter> {
   // Workaround for https://github.com/Microsoft/TypeScript/issues/3841.
   // Doing this makes our implements declaration work.
   "constructor": typeof MongooseAdapter;
 
-  public models: {[modelName: string]: mongoose.Model<any>};
-  public inflector: typeof pluralize;
-  public idGenerator: (doc: mongoose.Document) => mongodb.ObjectID
+  // Some precomputed name mappings that we use throughout.
+  // Note: a type name here can be a subtype (in the sense of Query.type,
+  // or what goes in a typePath), even though subtypes aren't rendered in
+  // the json-api `type` key. We need these mappings to read discriminator key
+  // values correctly, and it's also just more convenient for users to pass in
+  // models keyed by model name b/c that's how mongoose stores them.
+  protected typeNamesToModelNames: { [typeName: string]: string | undefined };
+  protected modelNamesToTypeNames: { [modelName: string]: string | undefined };
 
-  constructor(models?: {[modelName: string]: mongoose.Model<any>}, inflector?, idGenerator?) {
-    this.models = models || <{[modelName: string]: mongoose.Model<any>}>(<any>mongoose).models;
-    this.inflector = inflector || pluralize;
-    this.idGenerator = idGenerator;
+  constructor(
+    protected models: { [modelName: string]: Model<any> } = (mongoose as any).models,
+    protected toTypeName: (modelName: string) => string = getTypeName,
+    protected idGenerator?: ((doc: Document) => mongodb.ObjectID)
+  ) {
+    this.typeNamesToModelNames = {};
+    this.modelNamesToTypeNames = {};
+
+    for(const modelName of Object.keys(models)) {
+      const typeName = toTypeName(modelName);
+      this.typeNamesToModelNames[typeName] = modelName;
+      this.modelNamesToTypeNames[modelName] = typeName;
+    }
+  }
+
+  // Partially applied versions of the static/utility methods to simplify things.
+  docToResource(doc: Document, fields?: object) {
+    return docToResource(this.models, this.modelNamesToTypeNames, doc, fields);
+  }
+
+  docsToResourceData(docs: null | Document | Document[], isPlural: boolean, fields?: object) {
+    return this.constructor.docsToResourceData(
+      this.models,
+      this.modelNamesToTypeNames,
+      docs,
+      isPlural,
+      fields
+    );
+  }
+
+  getTypePath(model: Model<any>) {
+    return getTypePath(model, this.modelNamesToTypeNames);
   }
 
   /**
-   * Returns a Promise for an array of two items: the primary resources (either
-   * a single Resource or a Collection) and the included resources, as an array.
+   * Returns a Promise for an array of 3 items: the primary resources (either
+   * a single Resource or a Collection); the included resources, as an array;
+   * and the size of the full collection, if the primary resources represent
+   * a paginated view of some collection.
    *
    * Note: The correct behavior if idOrIds is an empty array is to return no
    * documents, as happens below. If it's undefined, though, we're not filtering
    * by id and should return all documents.
    */
-  find(type, idOrIds: string | string[] | undefined, fields, sorts, filters, includePaths, offset, limit) {
-    const model = this.getModel(this.constructor.getModelName(type));
-    const [mode, idQuery] = this.constructor.getIdQueryType(idOrIds);
-    const pluralizer = this.inflector.plural;
-    const isPaginating = typeof idOrIds !== 'string' && 
-      (typeof offset !== 'undefined' || typeof limit !== 'undefined');
-    const isFiltering = typeof filters === "object" && !Array.isArray(filters);
+  async find(query: FindQuery) {
+    const {
+      type,
+      populates: includePaths,
+      select: fields,
+      sort: sorts,
+      offset,
+      limit,
+      isSingular: singular
+    } = query;
 
-    let primaryDocumentsPromise, includedResourcesPromise = Q(null);
+    const mode = singular ? "findOne" : "find";
+    const filters = query.getFilters();
+    const mongofiedFilters = util.toMongoCriteria(filters);
+    const model = this.getModel(type);
 
-    const queryBuilder = mode === 'findOne' // ternary is a hack for TS compiler
-      ? model[mode](idQuery)
-      : model[mode](idQuery);
+    this.constructor.assertIdsValid(filters, singular);
 
-    // a promised query result that counts how many results we'd have if 
+    const isPaginating =
+      mode !== "findOne" &&
+      (typeof offset !== "undefined" || typeof limit !== "undefined");
+
+    let primaryDocumentsPromise: Promise<Document | Document[]>,
+      includedResourcesPromise: Promise<Resource[] | undefined>;
+
+    const queryBuilder =
+      mode === "findOne" // ternary is a hack for TS compiler
+        ? model[mode](mongofiedFilters)
+        : model[mode](mongofiedFilters);
+
+    // a promised query result that counts how many results we'd have if
     // we weren't paginating. this just resolves to null when we aren't paginating.
-    const collectionSizePromise = isPaginating 
-      ? model.count(isFiltering ? filters : undefined).exec()
+    const collectionSizePromise = isPaginating
+      ? model.count(mongofiedFilters).exec()
       : Promise.resolve(null);
 
     // do sorting
     if(Array.isArray(sorts)) {
-      queryBuilder.sort(sorts.join(" "));
-    }
+      const geoDistanceSort = sorts.find((it): it is ExpressionSort => {
+        const exp = (it as ExpressionSort).expression;
+        return exp && exp.operator === 'geoDistance';
+      });
 
-    // filter out invalid records with simple fields equality.
-    // note that there's a non-trivial risk of sql-like injection here.
-    // we're mostly protected by the fact that we're treating the filter's
-    // value as a single string, though, and not parsing as JSON.
-    if(isFiltering) {
-      queryBuilder.where(filters);
+      if(geoDistanceSort) {
+        if(sorts.length !== 1) {
+          throw Errors.invalidQueryParamValue({
+            detail: `Cannot combine geoDistance sorts with other sorts.`,
+            source: { parameter: "sort" }
+          });
+        }
+
+        queryBuilder.near(geoDistanceSort.expression.args[0].value, {
+          center: {
+            type: "Point", coordinates: geoDistanceSort.expression.args[1]
+          },
+          maxDistance: 4503599627370496,
+          spherical: true
+        })
+      }
+
+      else {
+        queryBuilder.sort(
+          sorts.map(it => {
+            if(!("field" in it)) {
+              throw new Error("Got unsupported expression sort field; shouldn't happen.");
+            }
+
+            return (it.direction === 'DESC' ? '-' : '') + it.field;
+          }).join(" ")
+        );
+      }
     }
 
     if(offset) {
@@ -82,19 +174,16 @@ export default class MongooseAdapter implements Adapter<typeof MongooseAdapter> 
     }
 
     // in an ideal world, we'd use mongoose here to filter the fields before
-    // querying. But, because the fields to filter can be scoped by type and
-    // we don't always know about a document's type until after query (becuase
-    // of discriminator keys), and because filtering out fields can really
-    // complicate population for includes, we don't yet filter at query time but
-    // instead just hide filtered fields in @docToResource. There is a more-
-    // efficient way to do this down the road, though--something like taking the
-    // provided fields and expanding them just enough (by looking at the type
-    // heirarachy and the relationship paths) to make sure that we're not going
-    // to run into any of the problems outlined above, while still querying for
-    // less data than we would without any fields restriction. For reference, the
-    // code for safely using the user's `fields` input, by putting them into a
-    // mongoose `.select()` object so that the user can't prefix a field with a
-    // minus on input to affect the query, is below.
+    // querying. But, because filtering out fields can really complicate population
+    // for (nested) includes, we don't yet filter at query time but instead just
+    // hide filtered fields in @docToResource. There is a more-efficient way to
+    // do this down the road, though -- something like taking the provided fields
+    // and expanding them just enough (by looking at the relationship paths) to
+    // make sure that we have the relationship fields to populate --while still
+    // querying for less data than we would without any fields restriction.
+    // For reference, the code for safely using the user's `fields` input, by
+    // putting them into a mongoose `.select()` object so that the user can't
+    // prefix a field with a minus on input to affect the query, is below.
     // Reference: http://mongoosejs.com/docs/api.html#query_Query-select.
     // let arrToSelectObject = (prev, curr) => { prev[curr] = 1; return prev; };
     // for(let type in fields) {
@@ -103,21 +192,22 @@ export default class MongooseAdapter implements Adapter<typeof MongooseAdapter> 
 
     // support includes, but only a level deep for now (recursive includes,
     // especially if done in an efficient way query wise, are a pain in the ass).
-    if(includePaths) {
+    if(includePaths && includePaths.length > 0) {
       const populatedPaths: string[] = [];
-      const refPaths = util.getReferencePaths(model);
+      const refPaths = getReferencePaths(model);
 
-      includePaths = includePaths.map((it) => it.split("."));
-      includePaths.forEach((pathParts) => {
+      includePaths.map((it) => it.split(".")).forEach((pathParts) => {
         // first, check that the include path is valid.
-        if(!arrayContains(refPaths, pathParts[0])) {
-          const title = "Invalid include path.";
-          const detail = `Resources of type "${type}" don't have a(n) "${pathParts[0]}" relationship.`;
-          throw new APIError(400, undefined, title, detail);
+        if(!refPaths.includes(pathParts[0])) {
+          throw Errors.invalidIncludePath({
+            detail: `Resources of type "${type}" don't have a(n) "${pathParts[0]}" relationship.`
+          });
         }
 
         if(pathParts.length > 1) {
-          throw new APIError(501, undefined, "Multi-level include paths aren't yet supported.");
+          throw Errors.unsupportedIncludePath({
+            detail: `Multi-level include paths like ${pathParts.join('.')} aren't yet supported.`
+          });
         }
 
         // Finally, do the population
@@ -125,415 +215,525 @@ export default class MongooseAdapter implements Adapter<typeof MongooseAdapter> 
         queryBuilder.populate(pathParts[0]);
       });
 
-      const includedResources: Resource[] = [];
-      primaryDocumentsPromise = Q(queryBuilder.exec()).then((docs) => {
-        forEachArrayOrVal(docs, (doc) => {
-          // There's no gaurantee that the doc (or every doc) was found
-          // and we can't populate paths on a non-existent doc.
-          if(!doc) return;
+      let includedResources: Resource[] = [];
+      primaryDocumentsPromise = Promise.resolve(queryBuilder.exec()).then((docOrDocs) => {
+        includedResources =
+          objectValues(
+            Data.fromJSON(docOrDocs) // fromJSON to handle docOrDocs === null.
+              .flatMap<Resource>((doc) => {
+                return Data.of(populatedPaths).flatMap((path) => {
+                  // Handle case that doc[path], which should hold id(s) of the
+                  // referenced documents, is undefined b/c the relationship isn't set.
+                  return typeof doc[path] === 'undefined'
+                    ? Data.empty as Data<Resource>
+                    : Data.fromJSON(doc[path]).map(docAtPath => {
+                        return this.docToResource(docAtPath, fields);
+                      });
+                });
+              })
+              .values
+              // Remove duplicates
+              .reduce((acc, resource) => {
+                acc[`${resource.type}/${resource.id}`] = resource;
+                return acc;
+              }, {} as { [id: string]: Resource })
+          );
 
-          populatedPaths.forEach((path) => {
-            // if it's a toOne relationship, doc[path] will be a doc or undefined;
-            // if it's a toMany relationship, we have an array (or undefined).
-            const refDocs = Array.isArray(doc[path]) ? doc[path] : [doc[path]];
-            refDocs.forEach((it) => {
-              // only include if it's not undefined.
-              if(it) {
-                includedResources.push(
-                  this.constructor.docToResource(it, pluralizer, fields)
-                );
-              }
-            });
-          });
-        });
-
-        return docs;
+        return docOrDocs;
       });
 
-      includedResourcesPromise = primaryDocumentsPromise.then(() =>
-        new Collection(includedResources)
-      );
+      includedResourcesPromise =
+        primaryDocumentsPromise.then(() => includedResources);
     }
 
     else {
-      primaryDocumentsPromise = Q(queryBuilder.exec());
+      primaryDocumentsPromise = Promise.resolve(queryBuilder.exec());
+      includedResourcesPromise = Promise.resolve(undefined);
     }
 
-    return Q.all([primaryDocumentsPromise.then((it) => {
-      const makeCollection = !idOrIds || Array.isArray(idOrIds) ? true : false;
-      return this.constructor.docsToResourceOrCollection(it, makeCollection, pluralizer, fields);
-    }), includedResourcesPromise, collectionSizePromise]).catch(util.errorHandler);
+    return Promise.all([
+      primaryDocumentsPromise.then((it) => {
+        return this.docsToResourceData(it, mode === 'find', fields);
+      }),
+      includedResourcesPromise,
+      collectionSizePromise
+    ])/*.then(([primary, included, collectionSize]) => {
+      return [primary, { included, meta: { collectionSize } }];
+    })*/.catch(util.errorHandler);
   }
 
   /**
    * Returns a Promise that fulfills with the created Resource. The Promise
    * may also reject with an error if creation failed or was unsupported.
-   *
-   * @param {string} parentType - All the resources to be created must be this
-   *   type or be sub-types of it.
-   * @param {(Resource|Collection)} resourceOrCollection - The resource or
-   *   collection of resources to create.
    */
-  create(parentType, resourceOrCollection) {
-    const resourcesByType = groupResourcesByType(resourceOrCollection);
+  async create(query: CreateQuery) {
+    const { records: resourceData } = query;
+    const getSmallestSubType = (it: ResourceWithTypePath) => it.typePath[0];
+    const setIdWithGenerator =
+      typeof this.idGenerator === "function" &&
+        ((doc) => { doc._id = (this.idGenerator as Function)(doc); });
 
-    // Note: creating the resources as we do below means that we do one
-    // query for each type, as opposed to only one query for all of the
-    // documents. That's unfortunately much slower, but it ensures that
-    // mongoose runs all the user's hooks.
-    const creationPromises: Array<Promise<mongoose.Document[]>> = [];
-    const setIdWithGenerator = (doc) => { doc._id = this.idGenerator(doc); };
-    for(const type in resourcesByType) {
-      const model = this.getModel(this.constructor.getModelName(type));
-      const resources: Resource[] = resourcesByType[type];
-      const docObjects = resources.map(util.resourceToDocObject);
+    const resourcesByParentType = partition('type', resourceData);
+    const creationPromises = Object.keys(resourcesByParentType).map(type => {
+      const model = this.getModel(type);
+      const discriminatorKey = getDiscriminatorKey(model);
 
-      if(typeof this.idGenerator === "function") {
-        forEachArrayOrVal(docObjects, setIdWithGenerator);
+      const resources = resourcesByParentType[type];
+      const docObjects = resources.map((resource) => {
+        const finalModel = this.getModel(getSmallestSubType(resource));
+        const forbiddenKeys = getMetaKeys(finalModel);
+
+        if(forbiddenKeys.some(k => k in resource.attrs || k in resource.relationships)) {
+          throw Errors.illegalFieldName();
+        }
+
+        return util.resourceToDocObject(resource, (typePath: string[]) => {
+          if(typePath.length === 1) {
+            return {};
+          }
+
+          const smallestSubType = getSmallestSubType(resource);
+          if(!discriminatorKey || !this.getModel(smallestSubType)) {
+            throw new Error("Unexpected model name. Should've been caught earlier.");
+          }
+
+          return { [discriminatorKey]: this.typeNamesToModelNames[smallestSubType] };
+        });
+      });
+
+      if(setIdWithGenerator) {
+        docObjects.forEach(setIdWithGenerator);
       }
 
-      creationPromises.push(model.create(docObjects));
-    }
+      return model.create(docObjects) as Promise<mongoose.Document[]>;
+    });
 
     return Promise.all(creationPromises).then((docArrays) => {
-      const makeCollection = resourceOrCollection instanceof Collection;
+      const makeCollection = !resourceData.isSingular;
       const finalDocs = docArrays.reduce((a, b) => a.concat(b), []);
       // TODO: why isn't fields passed here, but is only passed in find...
       // is that appropriate? (Probably... check spec, and common sense.)
-      return this.constructor.docsToResourceOrCollection(
-        finalDocs, makeCollection, this.inflector.plural
-      );
+      return this.docsToResourceData(finalDocs, makeCollection);
     }).catch(util.errorHandler);
   }
 
   /**
-   * @param {string} parentType - All the resources to be created must be this
-   *   type or be sub-types of it.
-   * @param {Object} resourceOrCollection - The changed Resource or Collection
-   *   of resources. Should only have the fields that are changed.
    */
-  update(parentType, resourceOrCollection) {
-    // It'd be faster to bypass Mongoose Document creation & just have mongoose
-    // send a findAndUpdate command directly to mongo, but we want Mongoose's
-    // standard validation and lifecycle hooks, and so we have to find first.
-    // Note that, starting in Mongoose 4, we'll be able to run the validations
-    // on update, which should be enough, so we won't need to find first.
-    // https://github.com/Automattic/mongoose/issues/860
-    // TODO: Fix the above. Using update also provides a better concurrency model.
-    // Either that, or we need to do add a where version = ... between the find
-    // and the update to make sure we don't have conflicting changes.
-    const model = this.getModel(this.constructor.getModelName(parentType));
-    const singular = this.inflector.singular;
-    const plural = this.inflector.plural;
+  async update(query: UpdateQuery) {
+    const { type: parentType, patch } = query;
+    const parentModel = this.getModel(parentType);
 
-    // Set up some data structures based on resourcesOrCollection
-    const resourceTypes: string[] = [];
-    const changeSets = {};
+    const prefetchedDocs = patch.map(it => it.adapterExtra).values.filter(it => !!it);
+    const getOIdAsString = R.pipe<Record<"_id", mongodb.ObjectID>, mongodb.ObjectID, string>(R.prop('_id'), String);
 
-    const idOrIds = mapResources(resourceOrCollection, (it: ResourceWithId) => {
-      changeSets[it.id] = it;
-      resourceTypes.push(it.type);
-      return it.id;
+    const docIdsToFetch = [...setDifference(
+      patch.map(R.prop('id')).values,
+      prefetchedDocs.map(getOIdAsString),
+    )];
+
+    // We only execute this if there's a non-zero number of docs still to fetch.
+    const remainingDocsQuery =
+      parentModel.find({ _id: { $in: docIdsToFetch } }).lean();
+
+    const docsToUpdate = docIdsToFetch.length === 0
+      ? prefetchedDocs
+      : [...prefetchedDocs, ...(await remainingDocsQuery.exec() as any)];
+
+    // We can store just by id, because we know all the types will be
+    // the same :). (They'll all be the root type of `UpdateQuery.type`.)
+    const docsToUpdateById = docsToUpdate.reduce(reduceToObject(getOIdAsString), {});
+
+    const updateOpts = {
+      new: true,
+      // we validate the full doc in memory below,
+      // so no need to run the validators again on update
+      runValidators: false,
+      upsert: false,
+      // discard updates to fields not recognized in the schema.
+      // TODO: set this to 'throw' instead?
+      strict: true
+    };
+
+    const singleDocUpdateQueries = await patch.mapAsync(async (resourceUpdate) => {
+      // tslint:disable-next-line no-shadowed-variable
+      const Model = this.getModel(resourceUpdate.typePath[0]);
+      const versionKey = getVersionKey(Model);
+
+      if(!Model) {
+        throw new Error("Unknown model name.");
+      }
+
+      const existingDoc = docsToUpdateById[resourceUpdate.id];
+      const changeSet = util.resourceToDocObject(resourceUpdate);
+
+      // Doc couldn't be found when we searched for it above.
+      if(!existingDoc) {
+        throw Errors.genericNotFound({
+          detail: `First missing resource was (${resourceUpdate.type}, ${resourceUpdate.id}).`
+        });
+      }
+
+      const forbiddenKeys = getMetaKeys(Model);
+      if(forbiddenKeys.some(k => k in changeSet)) {
+        throw Errors.illegalFieldName();
+      }
+
+      // Construct an in-memory document for this resource and
+      // apply the updates to it. Using .set ensures that setters get run, etc.
+      const updatedDoc = Model.hydrate(existingDoc).set(changeSet);
+
+      // Test that the doc is valid, using special
+      // mongo error handling function if it isn't.
+      try { await updatedDoc.validate(); } catch(e) { util.errorHandler(e) }
+
+      // Then, we can't .save() because that won't let us add a criteria
+      // on the version key. So, we construct an update based on the modified
+      // paths of the updatedDoc (again, we don't just use the changeSet because
+      // we want the result of applying setters) and findOneAndUpdate that.
+      // Note: even if we were hydrating the model with an empty doc, we'd want
+      // to call modifiedPaths to exclude keys for which the model set a default
+      const modifiedPaths = updatedDoc.modifiedPaths();
+      const updatedDocObject = updatedDoc.toObject();
+
+      const finalUpdate = modifiedPaths.reduce((acc, key) => {
+        acc[key] = updatedDocObject[key];
+        return acc;
+      }, {
+        // Increment the version key so other updates based on reads
+        // of an old version fail (with the version check below).
+        "$inc": { [versionKey]: 1 }
+      });
+
+      const criteria = {
+        _id: resourceUpdate.id,
+        [versionKey]: existingDoc[versionKey]
+      };
+
+      // Note: the query below has to be wrapped in an array (or something)
+      // so that it's not run automatically (because we're in an async function
+      // and query is a thenable, so js would await the query). We don't want
+      // it to run because we only wan to run the queries if all docs'
+      // validation succeeds.
+      return [parentModel.findOneAndUpdate(criteria, finalUpdate, updateOpts)];
     });
 
-    const [mode, idQuery] = this.constructor.getIdQueryType(idOrIds);
-    const queryBuilder = mode === 'findOne' // ternary is a hack for TS compiler
-      ? model[mode](idQuery)
-      : model[mode](idQuery);
+    // Don't run ANY of the queries unless all could be built successfully
+    // (I.e., every changeset was valid).
+    return singleDocUpdateQueries.flatMapAsync(async ([docUpdateQuery]) => {
+      const doc = await (docUpdateQuery.exec().catch(util.errorHandler) as Promise<mongoose.Document>);
 
-    return Q(queryBuilder.exec()).then((docs) => {
-      const successfulSavesPromises: Promise<mongoose.Document>[] = [];
-
-      // if some ids were invalid/deleted/not found, we can't let *any* update
-      // succeed. this is the beginning of our simulation of transactions.
-      // There are two types of invalid cases here: we looked up one or more
-      // docs and got none back (i.e. docs === null) or we looked up an array of
-      // docs and got back docs that were missing some requested ids.
-      if(docs === null) {
-        throw new APIError(404, undefined, "No matching resource found.");
-      }
-      else {
-        const idOrIdsAsArray = Array.isArray(idOrIds) ? idOrIds : [idOrIds];
-        const docIdOrIdsAsArray = Array.isArray(docs) ? docs.map(it => it.id) : [docs.id];
-
-        if(!arrayValuesMatch(idOrIdsAsArray, docIdOrIdsAsArray)) {
-          const title = "Some of the resources you're trying to update could not be found.";
-          throw new APIError(404, undefined, title);
-        }
+      if(!doc) {
+        throw Errors.occFail();
       }
 
-      forEachArrayOrVal(docs, (currDoc: mongoose.Document & { id: any}) => {
-        const newResource = changeSets[currDoc.id];
-
-        // Allowing the type to change is a bit of a pain. If the type's
-        // changed, it means the mongoose Model representing the doc must be
-        // different too. So we have to get the data from the old doc with
-        // .toObject(), change its discriminator, and then create an instance
-        // of the new model with that data. We also have to mark that new
-        // instance as not representing a new document, so that mongoose will
-        // do an update query rather than a save. Finally, we have to do all
-        // this before updating other attributes, so that they're correctly
-        // marked as modified when changed.
-        const currDocModel = (<mongoose.Model<any>>currDoc.constructor);
-        const currentModelName = currDocModel.modelName;
-        const newModelName = this.constructor.getModelName(newResource.type, singular);
-        if(currentModelName !== newModelName) {
-          const newDoc = currDoc.toObject({virtuals: true, getters: true});
-          const NewModelConstructor = this.getModel(newModelName);
-
-          // TODO: is schema.options.discriminatorKey always defined?
-          // Is this a/the most public interface through which we can access it?
-          newDoc[<string>(<any>currDocModel.schema).options.discriminatorKey] = newModelName;
-
-          // replace the currDoc with our new creation.
-          currDoc = new NewModelConstructor(newDoc);
-          currDoc.isNew = false;
-        }
-
-        // update all attributes and links provided, ignoring type/meta/id.
-        currDoc.set(util.resourceToDocObject(newResource));
-
-        successfulSavesPromises.push(currDoc.save());
-      });
-
-      return Promise.all(successfulSavesPromises);
-    }).then((docs) => {
-      const makeCollection = resourceOrCollection instanceof Collection;
-      return this.constructor.docsToResourceOrCollection(docs, makeCollection, plural);
-    }).catch(util.errorHandler);
+      return this.docsToResourceData(doc, false);
+    });
   }
 
-  delete(parentType, idOrIds) {
-    const model = this.getModel(this.constructor.getModelName(parentType));
-    const [mode, idQuery] = this.constructor.getIdQueryType(idOrIds);
-
-    if(!idOrIds) {
-      return Q.Promise((resolve, reject) => {
-        reject(new APIError(400, undefined, "You must specify some resources to delete"));
-      });
+  async delete(query: DeleteQuery) {
+    // Delete queries in theory let users delete resources that match arbitrary
+    // criteria. Atm, though, json-api only supports delete by id(s). And, below,
+    // we assume that ids are the only criteria in use, e.g., when we check the
+    // length of the result set. So, throw if that assumption doesn't hold.
+    if(!query.isSimpleIdQuery()) {
+      throw new Error("Unsupported delete query");
     }
 
-    const queryBuilder = mode === 'findOne' // ternary is a hack for TS compiler
-      ? model[mode](idQuery)
-      : model[mode](idQuery);
+    // On deletion, the library doesn't load + verify for us the subtypes of
+    // the resources the user is trying to delete (because that overhead isn't
+    // strictly necessary, unlike in the update case where it's needed to run
+    // the proper beforeSave). So, we have to do that ourselves first. If any
+    // types are invalid, we do no deletions and return 400.
+    const { type, isSingular: singular } = query;
+    const mode = singular ? 'findOne' : 'find';
+    const filters = query.getFilters();
+    const mongofiedFilters = util.toMongoCriteria(filters);
 
-    return Q(queryBuilder.exec()).then((docs) => {
-      if(!docs) throw new APIError(404, undefined, "No matching resource found.");
-      forEachArrayOrVal(docs, (it) => { it.remove(); });
-      return docs;
-    }).catch(util.errorHandler);
+    // Before we query to verify the subtypes, lets throw a targeted error
+    // message if we can identify that any of the ids are in an invalid format.
+    this.constructor.assertIdsValid(filters, singular);
+
+    // Ok, now, in order to verify that all ids of the right type, we need to
+    // query on the root model of `Query.type` (Query.type could be a subtype);
+    // if we query with the subtype model, ids that don't match the subtype will
+    // just not match any docs, and we won't be able to distinguish the
+    // "a doc is missing (404)" case from the "invalid type (400)" case.
+    const QueryTypeModel = this.getModel(type);
+    const baseModelName = QueryTypeModel.baseModelName || QueryTypeModel.modelName;
+    const BaseModel = this.getModel(<string>this.modelNamesToTypeNames[baseModelName]);
+
+    const queryBuilder = mode === 'findOne' // ternary is a hack for TS compiler
+      ? BaseModel[mode](mongofiedFilters)
+      : BaseModel[mode](mongofiedFilters);
+
+    const docsToDelete = await queryBuilder.exec().then((docOrDocsOrNull) => {
+      return Data.fromJSON<mongoose.Document>(docOrDocsOrNull);
+    }, util.errorHandler);
+
+    // Now, we verify all the type paths.
+    const hasTypePathThrough = (throughType: string, doc: Document) => {
+      return this.getTypePath(doc.constructor as Model<any>).includes(throughType);
+    }
+
+    if(!docsToDelete.every(R.partial(hasTypePathThrough, [type]))) {
+      throw Errors.invalidResourceType();
+    }
+
+    // Ok, having verified that all the resources are of the right (sub) type,
+    // we now want to check if no resources could be found and this is a single
+    // delete. In that case only, we return 404. (If some resources were missing
+    // and this is a bulk delete, we just delete the ones that we did find [as
+    // making the client retry is inconvenient, and the result of removing the
+    // missing ids and repeating the request would be the same].)
+    if(singular && docsToDelete.size === 0) {
+      throw Errors.genericNotFound();
+    }
+
+    // Finally, we can delete all the docs,
+    // calling .remove() to run the user's middleware.
+    docsToDelete.forEach(it => { it.remove(); });
+    return docsToDelete.map((it: mongoose.Document) => this.docToResource(it));
   }
 
   /**
    * Unlike update(), which would do full replacement of a to-many relationship
    * if new linkage was provided, this method adds the new linkage to the existing
    * relationship. It doesn't do a find-then-save, so some mongoose hooks may not
-   * run. But validation and the update query hooks will work if you're using
-   * Mongoose 4.0.
+   * run. But validation and the update query hooks will work.
    */
-  addToRelationship(type, id, relationshipPath, newLinkage) {
-    const model = this.getModel(this.constructor.getModelName(type));
+  async addToRelationship(query: AddToRelationshipQuery) {
+    const { type, id, relationshipName, linkage: newLinkage } = query;
+
+    // Below, we verify that the `type` key in linkage matches the `type` name
+    // of the root model that the relationship is supposed to hold.
+    //
+    // TODO: if the relationship is supposed to hold resources only of a
+    // specific subtype, validate that the linkage provided actually points to
+    // a document of that subtype (and actually points to an existing document
+    // at all!).
+    //
+    // TODO: update this to work like update(), where we find the doc and do an
+    // update on it so that save middleware can run and so we know the subtype
+    // of the owning document. Right now, we always run the parent model's
+    // findOneAndUpdate middleware, even if type/id points to a subtype resource.
+    // Likewise, we always read the supertype's schema when deciding which linkage
+    // types are allowed. That's bad too.
+    const model = this.getModel(type);
+    this.assertRelationshipLinkageTypeValid(model, relationshipName, newLinkage);
+
+    const options = { runValidators: true, context: "query" };
     const update = {
       $addToSet: {
-        [relationshipPath]: { $each: newLinkage.value.map(it => it.id)}
-      }
+        [relationshipName]: { $each: newLinkage.map(it => it.id) }
+      },
+      $inc: { [getVersionKey(model)]: 1 }
     };
-    const options = {runValidators: true};
 
-    return Q.ninvoke(model, "findOneAndUpdate", {"_id": id}, update, options)
+    return model.findOneAndUpdate({ "_id": id }, update, options).exec()
       .catch(util.errorHandler);
   }
 
-  removeFromRelationship(type, id, relationshipPath, linkageToRemove) {
-    const model = this.getModel(this.constructor.getModelName(type));
+  async removeFromRelationship(query: RemoveFromRelationshipQuery) {
+    const { type, id, relationshipName, linkage: linkageToRemove } = query;
+
+    // Below, we verify that the `type` key in linkage matches the `type` name
+    // of the root model that the relationship is supposed to hold.
+    //
+    // TODO: if the relationship is supposed to hold resources only of a
+    // specific subtype, validate that the linkage provided actually points to
+    // a document of that subtype (and actually points to an existing document
+    // at all).
+    //
+    // TODO: update this to work like update(), where we find the doc and do an
+    // update on it so that save middleware can run and so we know the subtype
+    // of the owning document. Right now, we always run the parent model's
+    // findOneAndUpdate middleware, even if type/id points to a subtype resource.
+    // Likewise, we always read the supertype's schema when deciding which linkage
+    // types are allowed. That's bad too.
+    const model = this.getModel(type);
+    this.assertRelationshipLinkageTypeValid(model, relationshipName, linkageToRemove);
+
     const update = {
       $pullAll: {
-        [relationshipPath]: linkageToRemove.value.map(it => it.id)
-      }
+        [relationshipName]: linkageToRemove.map(it => it.id)
+      },
+      $inc: { [getVersionKey(model)]: 1 }
     };
-    const options = {runValidators: true};
+    const options = { runValidators: true, context: "query" };
 
-    return Q.ninvoke(model, "findOneAndUpdate", {"_id": id}, update, options)
+    return model.findOneAndUpdate({"_id": id}, update, options).exec()
       .catch(util.errorHandler);
   }
 
-  getModel(modelName) {
-    if(!this.models[modelName]) {
-      // don't use an APIError here, since we don't want to
-      // show this internals-specific method to the user.
-      const err = new Error(`The model "${modelName}" has not been registered with the MongooseAdapter.`);
-      (<any>err).status = 404;
-      throw err;
+  /**
+   * Takes the json-api (i.e., parent-most) `type` and id for a set or resources,
+   * and returns the typePath for each resource. Must return subtypes earlier
+   * in the array, per typePath rules. Can also return extra data about the
+   * resource, which will be stored with it, for use optimizing future queries.
+   * Any resources that aren't found won't have an entry in the results object.
+   */
+  async getTypePaths(items: { type: string; id: string }[]) {
+    const itemsByType = partition("type", items);
+    const types = Object.keys(itemsByType);
+    const res = {};
+
+    for(const type of types) {
+      const theseItems = itemsByType[type];
+      const BaseModel = this.getModel(type);
+      const discriminatorKey = getDiscriminatorKey(BaseModel);
+
+      // When we have a .lean() doc straight from mongo, we can't
+      // just use .constructor to get the model. So we have to be
+      // a bit more clever.
+      const modelForLeanDoc = (it: any) =>
+        (discriminatorKey && it[discriminatorKey])
+          ? this.getModel(<string>this.modelNamesToTypeNames[it[discriminatorKey]])
+          : BaseModel;
+
+      // If we know the model has no subtypes,
+      // we can skip a query altogether and just return the type path.
+      if(!BaseModel.discriminators) {
+        res[type] = theseItems.reduce((acc, item) => {
+          acc[item.id] = { typePath: this.getTypePath(BaseModel) };
+          return acc;
+        }, {} as { [id: string]: TypeInfo });
+      }
+
+      else {
+        const docsPromise =
+          BaseModel.find({ _id: { $in: theseItems.map(it => it.id) }}).lean().exec();
+
+        res[type] = docsPromise.then(docs => {
+          return (docs as any[]).reduce((acc, doc) => {
+            acc[String(doc._id)] = {
+              typePath: this.getTypePath(modelForLeanDoc(doc)),
+              extra: doc
+            };
+            return acc;
+          }, {} as { [id: string]: TypeInfo });
+        })
+      }
     }
-    return this.models[modelName];
+
+    const values = await Promise.all(types.map(type => res[type]));
+    return types.reduce((acc, type, i) => {
+      acc[type] = values[i];
+      return acc;
+    }, {} as TypeIdMapOf<TypeInfo>);
   }
 
-  getTypesAllowedInCollection(parentType) {
-    const parentModel = this.getModel(
-      this.constructor.getModelName(parentType, this.inflector.singular)
-    );
-    return [parentType].concat(
-      this.constructor.getChildTypes(parentModel, this.inflector.plural)
-    );
+  getModel(typeName: string) {
+    const modelName = this.typeNamesToModelNames[typeName];
+    if(!modelName || !this.models[modelName]) {
+      // don't use an APIError here, since we don't want to
+      // show this internals-specific method to the user.
+      throw new Error(
+        `No model for type "${typeName}" is registered with the MongooseAdapter.`
+      );
+    }
+    return this.models[modelName];
   }
 
   /**
    * Return the paths that, for the provided type, must always must be filled
    * with relationship info, if they're present. Occassionally, a path might be
-   * optionally fillable w/ relationship info; this shouldn't return those paths.
+   * optionally fillable w/ relationship info [TODO: when? what does this mean?];
+   * this shouldn't return those paths.
    */
-  getRelationshipNames(type) {
-    const model = this.getModel(
-      this.constructor.getModelName(type, this.inflector.singular)
+  getRelationshipNames(typeName) {
+    const model = this.getModel(typeName);
+    return getReferencePaths(model);
+  }
+
+  doQuery(
+    query: CreateQuery | FindQuery | UpdateQuery | DeleteQuery |
+      AddToRelationshipQuery | RemoveFromRelationshipQuery
+  ) {
+    const method = (
+      (query instanceof CreateQuery && this.create) ||
+      (query instanceof FindQuery && this.find) ||
+      (query instanceof DeleteQuery && this.delete) ||
+      (query instanceof UpdateQuery && this.update) ||
+      (query instanceof AddToRelationshipQuery && this.addToRelationship) ||
+      (query instanceof RemoveFromRelationshipQuery && this.removeFromRelationship)
     );
-    return util.getReferencePaths(model);
+
+    if(!method) {
+      throw new Error("Unexpected query type.");
+    }
+
+    return method.call(this, query);
+  }
+
+  protected assertRelationshipLinkageTypeValid(
+    ownerModel: Model<any>,
+    relationshipName: string,
+    linkage: ResourceIdentifier[]
+  ) {
+    const rootTypeNameForRefModel = (() => {
+      try {
+        const refModelName = getReferencedModelName(ownerModel, relationshipName);
+        const refModelType = this.modelNamesToTypeNames[refModelName as string];
+        const refModel = this.getModel(refModelType as string);
+        return this.getTypePath(refModel).pop();
+      } catch(e) {
+        throw new Error(`Missing/invalid model name for relationship ${relationshipName}.`);
+      }
+    })();
+
+    if(!linkage.every(it => it.type === rootTypeNameForRefModel)) {
+      throw Errors.invalidLinkageType({
+        detail: `All linkage must have type: ${rootTypeNameForRefModel}.`
+      });
+    }
   }
 
   /**
-   * We want to always return a collection when the user is asking for something
-   * that's logically a Collection (even if it only has 1 item), and a Resource
-   * otherwise. But, because mongoose returns a single doc if you query for a
-   * one-item array of ids, and because we sometimes generate arrays (e.g. of
-   * promises for documents' successful creation) even when only creating/updating
-   * one document, just looking at whether docs is an array isn't enough to tell
-   * us whether to return a collection or not. And, in all these cases, we want
-   * to handle the possibility that the query returned no documents when we needed
-   * one, such that we must 404. This function centralizes all that logic.
+   * We want return a singular Data<Resource> when the result is conceptually
+   * singular, and a non-singular one otherwise. But, because mongoose returns
+   * a single doc if you query for a one-item array of ids, and because we
+   * sometimes generate arrays (e.g. of promises for documents' successful
+   * creation) even when only creating/updating one document, just looking at
+   * whether docs is an array isn't enough to tell us whether the result is
+   * singular. And, in all these cases, we want to handle the possibility that
+   * the query returned no documents when we needed one, such that we must 404.
+   * This function centralizes all that logic.
    *
-   * @param docs The docs to turn into a resource or collection
-   * @param makeCollection Whether we're making a collection.
-   * @param pluralizer An inflector function for setting the Resource's type
+   * @param {object} models A model name => model dictionary
+   * @param {object} modelNamesToTypeNames A model name => type name dictionary
+   * @param {null | mongoose.Document | mongoose.Document[]} docs The docs to
+   *   turn into a Data<Resource>
+   * @param {boolean} isPlural Whether the result is not conceptually singular.
+   * @param {object} fields
    */
-  static docsToResourceOrCollection(docs, makeCollection, pluralizer, fields?: object): Collection | Resource {
+  static docsToResourceData(
+    models: any,
+    modelNamesToTypeNames,
+    docs: null | mongoose.Document | mongoose.Document[],
+    isPlural: boolean,
+    fields?: object
+  ) {
     // if docs is an empty array and we're making a collection, that's ok.
     // but, if we're looking for a single doc, we must 404 if we didn't find any.
-    if(!docs || (!makeCollection && Array.isArray(docs) && docs.length === 0)) {
-      throw new APIError(404, undefined, "No matching resource found.");
+    if(!docs || (!isPlural && Array.isArray(docs) && docs.length === 0)) {
+      throw Errors.genericNotFound();
     }
 
-    docs = !Array.isArray(docs) ? [docs] : docs;
-    docs = docs.map((it) => this.docToResource(it, pluralizer, fields));
-    return makeCollection ? new Collection(docs) : docs[0];
+    const docsArray = !Array.isArray(docs) ? [docs] : docs;
+    const resources = docsArray.map((it) =>
+      docToResource(models, modelNamesToTypeNames, it, fields)
+    );
+
+    return isPlural ? Data.of<Resource>(resources) : Data.pure<Resource>(resources[0]);
   }
 
-  // Useful to have this as static for calling as a utility outside this class.
-  static docToResource(doc, pluralizer = pluralize.plural, fields?: object) {
-    const type = this.getType(doc.constructor.modelName, pluralizer);
-    const refPaths = util.getReferencePaths(doc.constructor);
-    const schemaOptions = doc.constructor.schema.options;
-
-    // Get and clean up attributes
-    // Note: we can't use the depopulate attribute because it doesn't just
-    // depopulate fields _inside_ the passed in doc, but can actually turn the
-    // doc itself into a string if the doc was originally gotten by population.
-    // That's stupid, and it breaks our include handling.
-    // Also, starting in 4.0, we won't need the delete versionKey line:
-    // https://github.com/Automattic/mongoose/issues/2675
-    let attrs = doc.toJSON({virtuals: true, getters: true});
-    delete attrs.id; // from the id virtual.
-    delete attrs._id;
-    delete attrs[schemaOptions.versionKey];
-    delete attrs[schemaOptions.discriminatorKey];
-
-    // Delete attributes that aren't in the included fields.
-    // TODO: Some virtuals could be expensive to compute, so, if field
-    // restrictions are in use, we shouldn't set {virtuals: true} above and,
-    // instead, we should read only the virtuals that are needed (by searching
-    // the schema to identify the virtual paths and then checking those against
-    // fields) and add them to newAttrs.
-    if(fields && fields[type]) {
-      const newAttrs = {};
-      fields[type].forEach((field) => {
-        if(attrs[field]) {
-          newAttrs[field] = attrs[field];
-        }
-      });
-      attrs = newAttrs;
-    }
-
-    // Build relationships
-    const relationships = {};
-    const getProp = (obj, part) => obj[part];
-
-    refPaths.forEach((path) => {
-      // skip if applicable
-      if(fields && fields[type] && !arrayContains(fields[type], path)) {
-        return;
-      }
-
-      // get value at the path w/ the reference, in both the json'd + full docs.
-      const pathParts = path.split(".");
-      let jsonValAtPath = pathParts.reduce(getProp, attrs);
-      const referencedType = this.getReferencedType(doc.constructor, path);
-
-      // delete the attribute, since we're moving it to relationships
-      deleteNested(path, attrs);
-
-      // Now, since the value wasn't excluded, we need to build its
-      // Relationship. Note: the value could still be null or an empty array.
-      // And, because of population, it could be a single document or array of
-      // documents, in addition to a single/array of ids. So, as is customary,
-      // we'll start by coercing it to an array no matter what, tracking
-      // whether to make it a non-array at the end, to simplify our code.
-      let isToOneRelationship = false;
-
-      if(!Array.isArray(jsonValAtPath)) {
-        jsonValAtPath = [jsonValAtPath];
-        isToOneRelationship = true;
-      }
-
-      let linkage: typeof Linkage.prototype.value = [];
-      jsonValAtPath.forEach((docOrIdOrNull) => {
-        let idOrNull;
-
-        // if it has an ._id key, it's a document.
-        if(docOrIdOrNull && docOrIdOrNull._id) {
-          idOrNull = String(docOrIdOrNull._id);
-        }
-
-        else {
-          // Even though we did toJSON(), id may be an ObjectId. (lame.)
-          idOrNull = docOrIdOrNull ? String(docOrIdOrNull) : null;
-        }
-
-        linkage.push(idOrNull ? {type: referencedType, id: idOrNull} : null);
-      });
-
-      // go back from an array if neccessary and save.
-      linkage = new Linkage(isToOneRelationship ? linkage[0] : linkage);
-      relationships[path] = new Relationship(linkage);
-    });
-
-    // finally, create the resource.
-    return new Resource(type, doc.id, attrs, relationships);
-  }
-
-  static getModelName(type, singularizer = pluralize.singular) {
-    const words = type.split("-");
-    words[words.length - 1] = singularizer(words[words.length - 1]);
-    return words.map((it) => it.charAt(0).toUpperCase() + it.slice(1)).join("");
-  }
-
-  // Get the json api type name for a model.
-  static getType(modelName, pluralizer = pluralize.plural) {
-    return pluralizer(modelName.replace(/([A-Z])/g, "\-$1").slice(1).toLowerCase());
-  }
-
-  static getReferencedType(model, path, pluralizer = pluralize.plural) {
-    return this.getType(util.getReferencedModelName(model, path), pluralizer);
-  }
-
-  static getChildTypes(model, pluralizer = pluralize.plural) {
-    if(!model.discriminators) return [];
-
-    return Object.keys(model.discriminators).map(it => this.getType(it, pluralizer));
-  }
-
-  static getStandardizedSchema(model, pluralizer = pluralize.plural) {
-    const schemaOptions = model.schema.options;
-    const versionKey = schemaOptions.versionKey;
-    const discriminatorKey = schemaOptions.discriminatorKey;
-    const virtuals = model.schema.virtuals;
+  static getStandardizedSchema(
+    model: mongoose.Model<any>,
+    pluralizer: typeof pluralize.plural = pluralize.plural.bind(pluralize)
+  ) {
+    const versionKey = getVersionKey(model);
+    const discriminatorKey = getDiscriminatorKey(model);
+    const virtuals = (model.schema as any).virtuals;
     const schemaFields: FieldDocumentation[] = [];
 
     const getFieldType = (path, schemaType) => {
@@ -541,65 +741,75 @@ export default class MongooseAdapter implements Adapter<typeof MongooseAdapter> 
         return new FieldTypeDocumentation("Id", false);
       }
 
-
       const typeOptions = schemaType.options.type;
       const holdsArray = Array.isArray(typeOptions);
 
       const baseType = holdsArray ? typeOptions[0].ref : typeOptions.name;
-      const refModelName = util.getReferencedModelName(model, path);
+      const refModelName = getReferencedModelName(model, path);
 
       return !refModelName ?
         new FieldTypeDocumentation(baseType, holdsArray) :
         new RelationshipTypeDocumentation(
-          holdsArray, refModelName, this.getType(refModelName, pluralizer)
+          holdsArray, refModelName, getTypeName(refModelName, pluralizer)
         );
 
     };
 
     model.schema.eachPath((name, type) => {
-      if(arrayContains([versionKey, discriminatorKey], name)) {
+      if([versionKey, discriminatorKey].includes(name)) {
         return;
       }
 
-      const fieldType = getFieldType(name, type);
-      name = (name === "_id") ? "id" : name;
-      const likelyAutoGenerated = name === "id" || (
-        fieldType.baseType === "Date" &&
-        /created|updated|modified/.test(name) &&
-        typeof type.options.default === "function"
-      );
+      // cast type to any for typescript because we're using
+      // an uncomfortable number of undocumented properties below.
+      const schemaType = type as any;
+      const fieldType = getFieldType(name, schemaType);
+      const publicName = name === "_id" ? "id" : name;
+
+      const likelyAutoGenerated =
+        publicName === "id" ||
+        (fieldType.baseType === "Date" &&
+          /created|updated|modified/.test(publicName) &&
+          typeof schemaType.options.default === "function");
 
       let defaultVal;
       if(likelyAutoGenerated) {
         defaultVal = "__AUTO__";
       }
 
-      else if(type.options.default && typeof type.options.default !== "function") {
-        defaultVal = type.options.default;
+      else if(schemaType.options.default && typeof schemaType.options.default !== "function") {
+        defaultVal = schemaType.options.default;
       }
 
       // find the "base type's" options (used below), in case
       // we have an array of values of the same type at this path.
-      const baseTypeOptions = Array.isArray(type.options.type) ? type.options.type[0] : type.options;
+      const baseTypeOptions = Array.isArray(schemaType.options.type)
+        ? schemaType.options.type[0]
+        : schemaType.options;
 
       // Add validation info
       const validationRules = {
-        required: !!type.options.required,
-        oneOf: baseTypeOptions.enum ? type.enumValues || (type.caster && type.caster.enumValues) : undefined,
-        max: type.options.max || undefined
+        required: !!schemaType.options.required,
+        oneOf: baseTypeOptions.enum
+          ? schemaType.enumValues ||
+            (schemaType.caster && schemaType.caster.enumValues)
+          : undefined,
+        max: schemaType.options.max || undefined
       };
 
-      type.validators.forEach((validatorObj) => {
+      schemaType.validators.forEach((validatorObj) => {
         Object.assign(validationRules, validatorObj.validator.JSONAPIDocumentation);
       });
 
-      schemaFields.push(new FieldDocumentation(
-        name,
-        fieldType,
-        validationRules,
-        this.toFriendlyName(name),
-        defaultVal
-      ));
+      schemaFields.push(
+        new FieldDocumentation(
+          publicName,
+          fieldType,
+          validationRules,
+          this.toFriendlyName(publicName),
+          defaultVal
+        )
+      );
     });
 
     for(const virtual in virtuals) {
@@ -611,12 +821,14 @@ export default class MongooseAdapter implements Adapter<typeof MongooseAdapter> 
       // for virtual properties, we can't infer type or validation rules at all,
       // so we add them with just a friendly name and leave the rest undefined.
       // The user is expected to override/set this in a resource type description.
-      schemaFields.push(new FieldDocumentation(
-        virtual,
-        undefined,
-        undefined,
-        this.toFriendlyName(virtual)
-      ));
+      schemaFields.push(
+        new FieldDocumentation(
+          virtual,
+          undefined,
+          undefined,
+          this.toFriendlyName(virtual)
+        )
+      );
     }
     return schemaFields;
   }
@@ -637,6 +849,7 @@ export default class MongooseAdapter implements Adapter<typeof MongooseAdapter> 
     const words: string[] = [];
     const wordsRe = /[A-Z]([A-Z]*(?![^A-Z])|[^A-Z]*)/g;
 
+    // tslint:disable-next-line no-conditional-assignment
     while((matches = wordsRe.exec(pascalCasedString)) !== null) {
       words.push(matches[0]);
     }
@@ -644,43 +857,81 @@ export default class MongooseAdapter implements Adapter<typeof MongooseAdapter> 
     return words.join(" ");
   }
 
-  /*
-  Note: adding overloads here creates more (false) compile errors than it fixes,
-  because of issues like https://github.com/Microsoft/TypeScript/issues/10523
-  static getIdQueryType(ids: string[]): ["find", {_id: {$in: string[]}}];
-  static getIdQueryType(id: string): ["findOne", {_id: string}];
-  static getIdQueryType(id: undefined): ["find", {}];
-  static getIdQueryType(): ["find", {}];*/
-  static getIdQueryType(idOrIds: string | string[] | undefined = undefined) {
-    type result =
-      ["findOne", {_id: string}] | // one resource
-      ["find", {_id: {$in: string[]}}] | // set of resources
-      ["find", {}]; // all resources in collection
+  /**
+   * This function checks that filter constraints on the id field are testing
+   * against the valid mongo id format, and throws a 400/404 error if not.
+   * The origin for it is that, by default, Mongoose would throw a CastError
+   * upon encountering an id string that it couldn't convert to an ObjectId,
+   * and we wanted to issue 404 errors in those cases (e.g., GET /people/1)
+   * rather than 500s, which is what the CastError got converted to.
+   *
+   * @param {AndExpression} filters A set of filter constraints to check.
+   * @param {boolean} isSingular Whether the query that the filters came from
+   *   is singular. Influences error message.
+   */
+  static assertIdsValid(filters: AndExpression, isSingular: boolean): void {
+    const idsArray = filters.args.reduce((acc, filter) => {
+      // We're only validating the RHS of binary operators where the left
+      // hand side is a reference to a field named id.
+      return isIdentifier(filter.args[0]) && filter.args[0].value === 'id'
+        ? acc.concat(filter.args[1] as any as (string | string[]))
+        : acc;
+    }, [] as string[]);
 
-    const [mode, idQuery, idsArray] = <result & { 2?: string[] }>(() => {
-      if(Array.isArray(idOrIds)) {
-        return ["find", {_id: {"$in": idOrIds}}, idOrIds];
-      }
-
-      else if(typeof idOrIds === "string") {
-        return ["findOne", {_id: idOrIds}, [idOrIds]];
-      }
-
-      else {
-        return ["find", {}, undefined];
-      }
-    })();
-
-    if (idsArray && !idsArray.every(this.idIsValid)) {
-      throw Array.isArray(idOrIds)
-        ? new APIError(400, undefined, "Invalid ID.")
-        : new APIError(404, undefined, "No matching resource found.", "Invalid ID.");
+    if(!idsArray.every(this.idIsValid)) {
+      throw isSingular
+        ? Errors.genericNotFound({ detail: "Invalid ID." })
+        : Errors.invalidId();
     }
-
-    return <result>[mode, idQuery];
   }
 
   static idIsValid(id) {
     return typeof id === "string" && /^[0-9a-fA-F]{24}$/.test(id);
   }
+
+  static supportedOperators: SupportedOperators = {
+    // These operators are treated with the library's built-in rules for
+    // whether they're binary and how to validate their arguments.
+    "and": { },
+    "or":  { },
+    'eq':  { },
+    'neq': { },
+    'ne':  { },
+    'in':  { },
+    'nin': { },
+    'lt':  { },
+    'gt':  { },
+    'lte': { },
+    'gte': { },
+
+    // Operators that other adapters really might not support, for which
+    // the library doesn't have built-in arg validation logic.
+    'geoDistance': {
+      isBinary: true,
+      // geoDistance produces a number not a bool,
+      // so it's used for sorting and not filtering
+      legalIn: ["sort"],
+      finalizeArgs(operators, operator, args) {
+        if(!isIdentifier(args[0])) {
+          throw new SyntaxError(
+            `"geoDistance" operator expects field reference as first argument.`
+          );
+        }
+
+        const isPoint = R.allPass([
+          Array.isArray, // tslint:disable-line no-unbound-method
+          R.pipe(R.length, R.equals(2)), // tslint:disable-line no-unbound-method
+          R.all(it => Number(it) === it)
+        ]);
+
+        if(!isPoint(args[1])) {
+          throw new SyntaxError(
+            `"geoDistance" operator expects [lng,lat] as second argument.`
+          );
+        }
+
+        return args;
+      }
+    }
+  };
 }
