@@ -1,16 +1,8 @@
-import APIError from "../../types/APIError";
+import { isPlainObject } from "../../util/misc";
+import * as Errors from "../../util/errors";
 
-export default function(body) {
-  return new Promise(function(resolve, reject) {
-    const ownProp = Object.prototype.hasOwnProperty;
-    const errMessage = "Request body is not a valid JSON API document.";
-
-    if(typeof body !== "object" || !ownProp.call(body, "data")) {
-      reject(new APIError(400, null, errMessage));
-    }
-
-    else {
-      resolve(undefined);
-    }
-  });
+export default async function(body) {
+  if(!isPlainObject(body) || !Object.prototype.hasOwnProperty.call(body, "data")) {
+    throw Errors.missingDataKey();
+  }
 }
