@@ -1,4 +1,4 @@
-import {expect} from "chai";
+import { expect } from "chai";
 import AgentPromise from "../../app/agent";
 
 describe("Fetching JSON Documentation", () => {
@@ -39,14 +39,10 @@ describe("Fetching HTML Documentation", () => {
       return Agent.request("GET", "/")
         .accept("text/*")
         .promise();
-    }, (e) => {
-      console.log(e); done(e);
-    }).then(response => {
+    }, done).then(response => {
       res = response;
       done();
-    }).catch((e) => {
-      console.log(e); done(e);
-    });
+    }).catch(done);
   });
 
   describe("Content Type", () => {
