@@ -11,7 +11,8 @@ export const VALID_ORG_RESOURCE_NO_ID = {
   "type": "organizations",
   "attributes": {
     "name": "Test Organization",
-    "modified": "2015-01-01T00:00:00.000Z"
+    "modified": "2015-01-01T00:00:00.000Z",
+    "echo": "Hello!"
   },
   "relationships": {
     "liaisons": {
@@ -21,7 +22,7 @@ export const VALID_ORG_RESOURCE_NO_ID = {
 };
 
 export const VALID_ORG_RESOURCE_NO_ID_EXTRA_MEMBER = {
-  ...VALID_ORG_RESOURCE_NO_ID,  
+  ...VALID_ORG_RESOURCE_NO_ID,
   "extraMember": true
 };
 
@@ -30,10 +31,38 @@ export const ORG_RESOURCE_CLIENT_ID = {
   "id": "53f54dd98d1e62ff12539db3"
 };
 
+export const ORG_RESOURCE_FALSEY_CLIENT_ID = {
+  ...VALID_ORG_RESOURCE_NO_ID,
+  "id": "0"
+};
+
+export const ORG_RESOURCE_FALSEY_CLIENT_ID_2 = {
+  ...VALID_ORG_RESOURCE_NO_ID,
+  "id": ""
+};
+
 export const VALID_SCHOOL_RESOURCE_NO_ID = {
-  "type": "schools",
+  "type": "organizations",
+  "meta": {
+    "types": ["organizations", "schools"]
+  },
   "attributes": {
-    "name": "Test School"
+    "name": "Test School",
+    "isCollege": false
+  }
+};
+
+export const VALID_SCHOOL_RESOURCE_NO_ID_EMPTY_PRINCIPAL_NO_LIAISONS = {
+  "type": "organizations",
+  "meta": {
+    "types": ["organizations", "schools"]
+  },
+  "attributes": {
+    "name": "Test School",
+    "isCollege": false
+  },
+  "relationships": {
+    "principal": { "data": null }
   }
 };
 
